Removed the checkNetworkConnection() function which only checked whether smartphone is connected to mobile network or wifi. It has been removed because:-
1.) It can be easily tricked by connecting smartphone to a wifi which is different from the
wifi to which the laptop is connected. This way although smartphone and
laptop are not connected the app will not display "Not Connected"
message.

2.) If the smartphone is hosting the hotspot then the smartphone must be connected to
mobile network otherwise checkNetworkConnection() will display "Not
Connected" message although both smartphone and laptop are connected. This is wrong as the app does not
require internet connection to connect to bottle server hence the app should run without
mobile network.

So inorder to fix this I created a "notConnected" boolean variable which
will keep track whether the smartphone is connected to bottle server or
not. If the smartphone is not connected then
        String response = HttpPost(urls[0])  --> line 197 in my code
will throw error. Hence "notConnected" is set as "true" in the catch
statement.
Finally an if-statement in the "onSensorChanged()" function will display
"Not Connected" message if the smartphone is not connected to bottle
server. "Not Connected" message will be displayed even if both
smartphone and laptop are connected to the same wifi but "sensor.py" has
not been started yet.

Also fixed the app from crashing if "Next" or "Back" button is pressed
immediately after launching the app.